file structure improved, dropdown moved to information item,separate folder for button templates added
