+++ client/src/App.jsx
@@ -1,7 +1,10 @@
 import {BrowserRouter} from "react-router-dom";
-import './scss/style.scss';
+import './scss/app.scss';
 import {RouterComponent} from "./components/routerComponent";
-import Header from "./components/common/Header";
+import Header from "./components/common/header/Header";
+import { useDispatch } from "react-redux";
+import { useEffect } from "react";
+import { getAuthStatus } from "./store/asyncActions";
 
 function App() {
     const dispatch = useDispatch();
