--- client/src/components/common/InteractionButton.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const InteractionButton = ({content = '', parentClass = ''}) => {
-  return (
-    <button className={parentClass + ' ' + 'interaction-btn btn'}>{content}</button>
-  )
-}
-
-export default InteractionButton;
