+++ client/src/components/common/buttonTemplates/CommunicationButton.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-
+import styles from './buttonTtemplates.module.scss'
 const CommunicationButton = ({iconURL, parentClass = ''}) => {
-
+  console.log();
   return (
-    <button className={parentClass + ' ' + 'communication-btn btn'}>
+    <button className={parentClass + ' ' + styles['communication-btn']}>
       {iconURL ? <img src={iconURL} alt="contact" /> : ''}
     </button>
   )
