+++ client/src/components/common/buttonTemplates/buttonTtemplates.module.scss
@@ -1,0 +1,38 @@
+@import "../../../scss/variables";
+@import "../../../scss/libs/normalize";
+.interaction-btn {
+  min-height: 63px;
+  width: 220px;
+  box-shadow: inset 0px 5px 5px rgba(0, 0, 0, 0.15);
+  border-radius: 7px;
+  font-size: 16px;
+  font-weight: 700;
+  letter-spacing: 0em;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: $lilac;
+  color: #FFFFFF;
+  transition: all 3ms ease-in-out;
+  &:hover {
+    filter: brightness(105%);
+  }
+  &:active {
+    filter: brightness(125%);
+  }
+}
+.communication-btn {
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: $lilac;
+  color: #FFFFFF;
+  transition: all 3ms ease-in-out;
+  &:hover {
+    filter: brightness(105%);
+  }
+  &:active {
+    filter: brightness(125%);
+  }
+}
