+++ client/src/pages/basket.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import '../scss/style.scss';
 
 function Basket(props) {
     return (
