+++ client/src/pages/home.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import '../scss/style.scss';
 
 function Home(props) {
     return (
