+++ client/src/pages/item.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import '../scss/style.scss';
 import {useParams} from "react-router-dom";
 
 function Item(props) {
