+++ client/src/pages/order.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import '../scss/style.scss';
 
 function Order(props) {
     return (
