+++ client/src/scss/_variables.scss
@@ -1,6 +1,5 @@
-//fonts
-@import url('https://fonts.googleapis.com/css2?family=PT+Sans:ital,wght@0,400;0,700;1,400;1,700&family=Roboto&family=Work+Sans:wght@300;400&display=swap');
 //colors
+$bg_color: #E5E5E5;
 $black: #333333;
 $lilac: #956D84;
 $green: #828282;
