+++ client/src/scss/app.scss
@@ -1,0 +1,23 @@
+@import "./libs//normalize";
+@import "variables";
+@import "./components/header";
+body {
+  background: $bg_color;
+  height: 100%;
+  line-height: 1;
+  font-size: 16px;
+  color: black;
+  font-family: 'Roboto';
+}
+.wrapper {
+  min-height: 100%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+}
+.container {
+  max-width: 1421px;
+  margin: 0 auto;
+  padding: 0px 15px;
+  box-sizing: content-box;
+}
