+++ client/src/scss/components/_footer.scss
@@ -1,0 +1,1 @@
+@import '../variables';
